refactor(frontend): a declared seam, and the first two regions through it
`panel.ts` is one class rendering every surface, so lifting a region out of it
needs an answer to "what may the region still reach?". This is that answer:
`panel-host.ts` declares `PanelHost`, the slice of the panel a region module may
touch, and `HomeKeeperPanel implements PanelHost`. Eleven members, sorted, each
one a deliberate widening — the alternative is passing the whole class as `any`
and finding out at runtime.

Two regions move through it, as free functions rather than sub-controllers:

- `panel-history.ts` — the completion groups a detail page shows, their markup,
  and the per-row delete/edit/move wiring, plus the three DOM helpers the detail
  pages and appliance editor build chrome from (`section`, `collapsibleSection`,
  `setIcon`), which the history rendering already shared.
- `panel-chips.ts` — the area / tag / managed / device chips and the one pass
  that makes every device chip clickable.

`taskChipsList` turned out to be a pure function of the task, and the card had a
character-for-character copy of it inline; the card now calls it, so the two
surfaces can no longer drift. `wireDeviceChips` needs no host at all — the chips
carry their device id in the DOM — so it doesn't take one.

Three hazards that outlived their region are now written down where the field
lives: `_liveHassEls` is push-only (its reset belongs to `_render`), an open
appliance's `.asset` is mutated in place and read back by `_submitAssetForm`, and
`_attachNotePreview` stays in `panel.ts` as the only preview constructor so
`_disposeAllPreviews` can still see them all.

No behaviour: 793 tests unchanged, and the bundles differ only by these renames.
